requirements.txt
dill==0.3.7
dominate==2.9.0
einops==0.7.0
imageio==2.33.0
lpips==0.1.4
matplotlib==3.5.0
mlflow==2.8.1
monai==0.9.1
moviepy==1.0.3
nibabel==4.0.2
nipype==1.8.6
numpy==1.26.2
omegaconf==2.3.0
pandas==2.1.3
Pillow==10.1.0
pynvml==11.5.0
scikit_learn==1.3.2
scipy==1.11.4
scikit-image
statsmodels==0.14.0
tensorboardX==2.6.2.2
tensorflow==2.15.0
torch==1.12.1
torchvision==0.13.1
tqdm==4.66.1
